tst_qtgraphicaleffects: make the test work prior to installation
Tests should be working both before and after `make install'.
Adjust this test so it can load QtGraphicalEffects either from the
source tree or the install tree.

This code already attempted to accomplish this:

  engine.addImportPath("../../src");

It didn't work, since the directory in the source tree is named
"effects" and not "QtGraphicalEffects".  It also would not work when
the build tree and source tree are different.

Change-Id: Ied964714d72e041caa937a4204668fe4f86da3f5