tst_qtgraphicaleffects: make the test work on devices
Make the test find the 'QtGraphicalEffects' import under
Qml2ImportsPath (QML 2.x) instead of ImportsPath (QML 1.x)
when the source code is not available.

This seems to have worked so far because the source code is always
available in the CI machines, but it fails on devices where is not.

Change-Id: I475ccd5431d5e7c3adf239b1158d7bb67e0bc042
Reviewed-by: Gunnar Sletta <gunnar.sletta@jollamobile.com>