Let sourceproxy auto-configure layers.
In the old SourceProxy implementation, layer.enabled = true that the
layer was used without a shader source in between. However, since
effects like Blur and DropShadow requires updates to sourceRect and
such, the effects would not work properly.

Since the usecase for layers is primarily to combine it with
layer.effect, we configure the layer.

[ChangeLog] When applying an effect to Item::layer.effect, the effect
will update the layer properties to make the effect work, such as
'smooth' to 'true' and changing 'sourceRect' to take
'transparentBorder' into account.

Change-Id: Idd68b025ade46c0b84142b2afb244730bed863d2
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>