Implement DropShadow using the new gaussian blur.
The 'fast' path has been discarded because it fails to deliver on
the fast promise. For radii less than 30-40 the normal blur is faster,
so little point

[ChangeLog] DropShadow uses the new gaussian blur
implementation. 'fast' no longer has any effect. The default values
have been updated so that 'layer.effect: DropShadow { }' works out of
the box.

Change-Id: I7e4ace4623613f0969a35e0c9bb873c18a158450
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>