Implement MaskedBlur in terms of the new GaussianBlur.
The mask code uses the fallback shader which is equivalent,
performance wise, to what we have in Qt 5.4, but in the interest of
keeping one codebase for all the blurring, we'll replace it anyway.

'transparentBorder' doesn't do anything anymore because it never
really worked to begin with and it doesn't make sense. See comment in
code.

'fast' codepath has been removed because it isn't really all that
fast.

[ChangeLog] MaskedBlur has a new blur implementation. The default
values of properties has been updated such that 'layer.effect:
MaskedBlur { maskSource: myMask }' works out of the box. The 'fast'
and 'transparentBorder' properties no longer have any effect.

Change-Id: Ibc05442914b563c50c95212fb0d4578f02990504
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>