Make relevant effects respect layer.smooth
Currently, source items that set smooth to false will have blurry
effects when using layer.effect. Fix that by setting the interpolation
property of each effect's SourceProxy.

Standalone effects seem to work fine already.

The patch adds a manual testcase: "smooth". This can be used to verify
that the functionality works in the future, as adding an auto test
is not really possible.

This isn't done for all effects, as it doesn't make sense for all of
them (blurs, for example).

Task-number: QTBUG-67382
Change-Id: Ic6f030090cd0579dd4259ebafc9c99a8b9940494
Reviewed-by: Andy Nichols <andy.nichols@qt.io>