Android: Fix autotest
First of all, we need to include a dummy.qml which imports the
graphical effects, otherwise it will be ignored by qmlimportscanner,
since all the QML code is generated programmatically.

Second of all, using the Qml2ImportPath to look up the imports is
not sufficient on Android, where the QML2_IMPORT_PATH environment
variable is used instead. So we just add code to check this first,
based on the order of precedence actually used by Qt.

Task-number: QTBUG-73574
Change-Id: I7b140dcac758ff37340fd7539f5cacb2982ff7aa
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>
Reviewed-by: BogDan Vatra <bogdan@kdab.com>