Fix graphical effects on images with no dimensions
When we pick the direct code path for a null image (if you
set its source to ""), the shader effect will try to sample
non-existent pixels and get black. Whereas if we use a proxy
for this, then it will create an empty source item the size of
the item instead.

This used to work by accident before
3c4c7ac6069a57b17199627742d8e2b32a3a31c9, because the interpolation
was always set to the wrong value, so interopOk would always be false and
we would always use a proxy. But after that was fixed, we started hitting
the direct code path and null images broke as a result.

The easy fix is just to go back to using the proxy item code path when there
is no actual input.

Note: This is only committed to 5.15, as the same issue is not visible in
Qt 6, when running on RHI.

Fixes: QTBUG-86486
Change-Id: I1b7927fa1a407f70703e8c4fa380c1c3582c0aaf
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>