Update commercial license headers
Updated header.COMMM to the files in tqtc-qtgraphicaleffects.
Examples, documentation files or tests are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotest use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4164
Change-Id: I8eed3e1c06dd0c742026bac476ba7832ba5c29da
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>